requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.19.2
anyio==4.3.0
assemblyai==0.22.0
async-timeout==4.0.3
attrs==23.2.0
blobfile==2.1.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
evaluate==0.4.1
exceptiongroup==1.2.0
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.45.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.25.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.0
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.3
hyperframe==6.0.1
idna==3.6
kiwisolver==1.4.5
lxml==4.9.4
matplotlib==3.8.3
multidict==6.0.5
multiprocess==0.70.16
numpy==1.24.4
oauthlib==3.2.2
openai==1.14.1
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip==24.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.8.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
setuptools==69.1.0
sgmllib3k==1.0.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
tiktoken==0.1.1
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
